add double type support

+++ IdeaProjects/Курс Java для начинающих 1.0/src/Operators4.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Operators4 {
     public static void main (String[] args) {
     Scanner scanner = new Scanner(System.in);
-    int x = scanner.nextInt();
-    int y = scanner.nextInt();
-    int z = scanner.nextInt();
+    double x = scanner.nextDouble();
+    double y = scanner.nextDouble();
+    double z = scanner.nextDouble();
     double avg = (x + y + z)/3;
     double halfAvg = avg/2;
     if (Math.floor(halfAvg) > 3)
